Support --output-format (and -O)
Allow specifying an output format when using an external downloader,
either through --output-format or -O, in preference to the extractor's
default. This is particularly helpful as -O mkv, to allow Matroska
output from ffmpeg in cases where .mp4 isn't a great answer, such as
while trying to play during the download.